Refactoring: Remove cronjob_role param - nowhere used
Also relocates cronjob related functions from shared lib to cronjob provider
