+++ easybib/providers/crontab.rb
@@ -31,7 +31,7 @@ action :create do
 
   create_crontab = deploy_crontab?(
     new_resource.instance_roles,
-    new_resource.cronjob_role
+    node['easybib_deploy']['cronjob_role']
   )
   if create_crontab
     cron = ::EasyBib::Cron.new(app, crontab_file)
@@ -88,4 +88,11 @@ action :delete do
 
   new_resource.updated_by_last_action(true)
 
+end
+
+def deploy_crontab?(instance_roles, cronjob_role)
+  return true if cronjob_role.nil?
+  return true if instance_roles.include?(cronjob_role)
+  Chef::Log.info('Instance is not in a cronjob role, skippings cronjob installs')
+  false
 end
