+++ easybib/resources/crontab.rb
@@ -5,5 +5,4 @@ default_action :create
 attribute :app, :default => ''
 attribute :crontab_file, :kind_of => String, :default => ''
 attribute :instance_roles, :default => []
-attribute :cronjob_role, :kind_of => String, :default => nil
 attribute :crontab_user, :kind_of => String, :default => 'www-data'
